requirements.txt
asgiref==3.8.1
cloudinary==1.42.1
django-cloudinary-storage==0.3.0
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
Django==4.2.6
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.1
sslcommerz-lib==1.0
whitenoise==6.8.2
